refactor : Entity pk Long타입 변경 및 코드 리팩토링

+++ src/main/java/com/ll/exam/sbb/answer/AnswerController.java
@@ -20,7 +20,7 @@ public class AnswerController {
     private final QuestionService questionService;
 
     @PostMapping("/create/{id}")
-    public String save(Model model, @PathVariable("id") int questionId, @Valid AnswerForm answerForm, BindingResult bindingResult) {
+    public String save(Model model, @PathVariable("id") Long questionId, @Valid AnswerForm answerForm, BindingResult bindingResult) {
         Question question = questionService.getQuestion(questionId);
 
         // 필수 입력 폼 유효성 검증
